Update CustomFieldRepositoryTrait.php
For parsing purposes of a lead in api, having it to show the visibility of a field can be very useful. 
e.g: i made this change locally to rebuild the forms and when posting to the server, i would only send the fields that could be publically visible and by the same parameter, use this to build in a form only those fields.